moved more helpers and updated paths
