+++ packages/client/src/Components/Mapping/helpers/hasPendingForces.js
