+++ packages/client/src/Components/Mapping/helpers/mousePosition.js
